Clean up and pick up consolidated wrappers and tests
