+++ packages/graphql-mocks/src/wrapper/index.ts
@@ -1,0 +1,3 @@
+export { spyWrapper } from './spy';
+export { logWrapper } from './log';
+export { stashStateWrapper, stashFor } from './stash-state';
