+++ packages/graphql-mocks/src/wrapper/spy.ts
