+++ packages/graphql-mocks/src/wrapper/stash-state.ts
