+++ packages/graphql-mocks/test/unit/graphql/handler.test.ts
@@ -3,7 +3,7 @@ import * as sinon from 'sinon';
 import { expect } from 'chai';
 import { GraphQLHandler } from '../../../src/graphql';
 import { ResolverMap } from '../../../src/types';
-import { spyWrapper } from '../../../src/spy';
+import { spyWrapper } from '../../../src/wrapper';
 import { embed } from '../../../src/resolver-map/embed';
 
 describe('graphql/hander', function () {
